# docs/swagger.yaml
basePath: /
...
info:
  contact:
    name: Fern Platform Team
    url: https://github.com/guidewire-oss/fern-platform/issues
  description: Unified test intelligence platform — aggregate, analyze and act on
    test results from any CI/CD pipeline.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/guidewire-oss/fern-platform
  title: Fern Platform API
  version: "1.0"
paths:
...
  /api/v1/summary/{projectId}/{seed}:
    get:
      description: Returns an aggregated test summary for a project, optionally grouped
        by one or more dimensions
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      - description: Test seed or run identifier
        in: path
        name: seed
        required: true
        type: string
      - collectionFormat: multi
        description: Dimensions to group results by (repeatable)
        in: query
        items:
          type: string
        name: group_by
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get test summary
      tags:
      - test-runs
...
securityDefinitions:
  BearerAuth:
    description: 'JWT bearer token. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...